plugins/medical-imaging/requirements.txt
mcp
torch
torchvision
pillow
numpy
flask
requests
pydicom
torchxrayvision
